+++ next-docs/components/sidebar/navigation.ts
@@ -21,10 +21,11 @@ const navigation = [
       { name: 'Icons', href: '/core/icons' },
       { name: 'Label', href: '/core/label' },
       { name: 'Loader', href: '/core/loader' },
-      { name: 'TextInput Group', href: '/core/textInputGroup' },
       { name: 'TextInput', href: '/core/textInput' },
+      { name: 'TextInput Group', href: '/core/textInputGroup' },
       { name: 'Typography', href: '/core/typography' },
       { name: 'Pagination', href: '/core/pagination' },
+      { name: 'Radio button', href: '/core/radiobutton' },
     ],
   },
   {