+++ next-docs/public/examples/avatarTW/Variants.tsx
@@ -3,7 +3,7 @@ import { Avatar } from '@heathmont/moon-core-tw';
 const Example = () => (
   <>
     <Avatar />
-    <Avatar name="JS" />
+    <Avatar name="md" />
     <Avatar imageUrl="https://www.fillmurray.com/200/200" />
   </>
 );