+++ next-docs/public/examples/radiobuttonTW/Disabled.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { RadioButton } from '@heathmont/moon-core-tw';
+
+const Example = () => (
+  <RadioButton
+    disabled
+    name="disabled"
+    label="Disabled radio button"
+    ariaLabel="Disabled radio button"
+    id="rb1"
+  />
+);
+export default Example;